pyproject.toml
[tool.poetry]
name = "open-nl2data-bench"
version = "0.1.0"
description = "OpenNL2Data-Bench: A production-grade benchmark framework for NL2Data systems (NL2SQL, NL2API, NL2Python, etc.)"
authors = ["PowerData Community <team@powerdata.org>"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/powerdata-ai/open-nl2data-bench"
repository = "https://github.com/powerdata-ai/open-nl2data-bench"
keywords = ["nl2data", "nl2sql", "nl2api", "nl2python", "benchmark", "chatbi", "data-agent", "llm"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Database",
]
packages = [{include = "onb"}]

[tool.poetry.dependencies]
python = "^3.10"
# Core dependencies
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
PyYAML = "^6.0.1"

# Database adapters
sqlalchemy = "^2.0.23"
pymysql = "^1.1.0"
psycopg2-binary = "^2.9.9"  # PostgreSQL adapter
clickhouse-driver = "^0.2.6"  # ClickHouse native driver
clickhouse-sqlalchemy = "^0.3.0"  # ClickHouse SQLAlchemy dialect
cryptography = "^41.0.7"  # Required by PyMySQL for SSL

# Data processing
pandas = "^2.1.4"
pyarrow = "^14.0.1"

# CLI framework
typer = {extras = ["all"], version = "^0.9.0"}
rich = "^13.7.0"

# HTTP client for SUT adapters
httpx = "^0.25.2"
aiohttp = "^3.9.1"

# Utilities
python-dotenv = "^1.0.0"
loguru = "^0.7.2"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"

# Code quality
mypy = "^1.7.1"
ruff = "^0.1.8"
black = "^23.12.0"
isort = "^5.13.2"

# Type stubs
pandas-stubs = "^2.1.4"
types-PyYAML = "^6.0.12"

[tool.poetry.group.data.dependencies]
# Data generation (for MVP-1+)
faker = "^20.1.0"

[tool.poetry.group.stress.dependencies]
# Load testing (for Phase 2)
locust = "^2.20.0"

[tool.poetry.scripts]
onb = "onb.cli.main:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=onb",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "slow: Tests that take a long time",
]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = [
    "pandas.*",
    "pyarrow.*",
    "faker.*",
    "locust.*",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py310"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true

[tool.coverage.run]
source = ["onb"]
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
